refactor: extract package name validation
