+++ Cargo.toml
@@ -43,6 +43,7 @@ pnpm-text-sanitize                     = { path = "pnpm/crates/text-sanitize" }
 pnpm-pack                              = { path = "pnpm/crates/pack" }
 pnpm-publish                           = { path = "pnpm/crates/publish" }
 pnpm-python-resolver                   = { path = "pnpm/crates/python-resolver" }
+pnpm-package-name                      = { path = "pnpm/crates/package-name" }
 pnpm-package-manifest                  = { path = "pnpm/crates/package-manifest" }
 pnpm-package-manager                   = { path = "pnpm/crates/package-manager" }
 pnpm-package-is-installable            = { path = "pnpm/crates/package-is-installable" }
