+++ pnpm/crates/cli/Cargo.toml
@@ -15,6 +15,7 @@ name = "pnpm"
 path = "src/bin/main.rs"
 
 [dependencies]
+pnpm-package-name                      = { workspace = true }
 cargo-util-schemas                     = { workspace = true }
 pnpm-auth-commands                     = { workspace = true }
 pnpm-cargo-resolver                    = { workspace = true }
