+++ pnpm/crates/lockfile-verification/Cargo.toml
@@ -11,11 +11,11 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
-pnpm-diagnostics                       = { workspace = true }
-pnpm-lockfile                          = { workspace = true }
-pnpm-reporter                          = { workspace = true }
-pnpm-resolving-parse-wanted-dependency = { workspace = true }
-pnpm-resolving-resolver-base           = { workspace = true }
+pnpm-package-name            = { workspace = true }
+pnpm-diagnostics             = { workspace = true }
+pnpm-lockfile                = { workspace = true }
+pnpm-reporter                = { workspace = true }
+pnpm-resolving-resolver-base = { workspace = true }
 
 chrono       = { workspace = true }
 derive_more  = { workspace = true }
