+++ pnpm/crates/lockfile-verification/src/verify_lockfile_resolutions.rs
@@ -17,10 +17,10 @@ use std::{collections::BTreeMap, path::Path, sync::Arc, time::Instant};
 
 use futures_util::{StreamExt, stream::FuturesUnordered};
 use pnpm_lockfile::{Lockfile, LockfileResolution, PkgName, is_git_hosted_tarball_url};
+use pnpm_package_name::is_valid_old_npm_package_name;
 use pnpm_reporter::{
     LockfileVerificationLog, LockfileVerificationMessage, LogEvent, LogLevel, Reporter,
 };
-use pnpm_resolving_parse_wanted_dependency::is_valid_old_npm_package_name;
 use pnpm_resolving_resolver_base::{
     ResolutionPolicyViolation, ResolutionVerification, ResolutionVerifier, VerifyCtx, VerifyFuture,
 };
