+++ pnpm/crates/package-manager/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
+pnpm-package-name                      = { workspace = true }
 pnpm-matcher                           = { workspace = true }
 rustc-hash                             = { workspace = true }
 pnpm-catalogs-config                   = { workspace = true }
