+++ pnpm/crates/package-name/src/lib.rs
@@ -1,33 +1,9 @@
-//! Inline port of the `validForOldPackages` branch of npm's
-//! [`validate-npm-package-name`](https://www.npmjs.com/package/validate-npm-package-name)
-//! (v7.0.2, the version pnpm pins in its workspace catalog). Pacquet only
-//! needs the boolean "is this still a usable package name?" answer at
-//! the call site in [`crate::parse_wanted_dependency`], so the
-//! warnings-vs-errors distinction and the per-rule error message strings
-//! that the upstream JS library returns are intentionally not modeled.
-//!
-//! Mirrors the JS implementation at
-//! `validate-npm-package-name/lib/index.js` (v7.0.2). The function
-//! returns `true` exactly when upstream's `validForOldPackages` would.
+//! npm package-name validation using the `validForOldPackages` rules.
 
-/// `true` when `name` would have an empty `errors` array under
-/// `validate-npm-package-name@7`, i.e. upstream's
-/// `validForOldPackages === true`.
-///
-/// The rules that flip this to `false` are, in order:
-///
-/// 1. empty string
-/// 2. starts with `.`
-/// 3. starts with `-`
-/// 4. starts with `_`
-/// 5. has leading or trailing ASCII whitespace
-/// 6. equals (case-insensitive) `node_modules` or `favicon.ico`
-/// 7. contains characters that aren't URL-safe in the `encodeURIComponent`
-///    sense, **except** the scoped-name shape `@user/pkg` where both
-///    halves are individually URL-safe and `pkg` does not start with `.`
-///
-/// The URL-safe character set matches JS's `encodeURIComponent`: ASCII
-/// letters and digits plus `- _ . ! ~ * ' ( )`.
+pub use is_valid_old_npm_package_name as is_valid_dependency_alias;
+
+/// Whether npm's `validate-npm-package-name` v7 accepts `name` for old packages.
+/// Warning-only names remain valid, including uppercase names and legacy lengths.
 #[must_use]
 pub fn is_valid_old_npm_package_name(name: &str) -> bool {
     if name.is_empty() {
@@ -46,10 +22,6 @@ pub fn is_valid_old_npm_package_name(name: &str) -> bool {
         return true;
     }
     if let Some((user, pkg)) = match_scoped(name) {
-        // v7 added the explicit leading-`.` reject on the pkg half
-        // inside the scoped branch; without it `@scope/.foo` would
-        // sneak through as URL-safe. Mirrors the new lines around
-        // `validate-npm-package-name/lib/index.js@7.0.2` L83-L85.
         if pkg.starts_with('.') {
             return false;
         }
@@ -95,3 +67,9 @@ fn match_scoped(name: &str) -> Option<(&str, &str)> {
     }
     Some((user, pkg))
 }
+
+#[cfg(test)]
+mod tests;
+
+#[cfg(test)]
+mod alias_tests;
