+++ pnpm/crates/resolving-deps-resolver/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
+pnpm-package-name                      = { workspace = true }
 tracing                                = { workspace = true }
 rustc-hash                             = { workspace = true }
 pnpm-catalogs-resolver                 = { workspace = true }
