+++ pnpm/crates/resolving-npm-resolver/src/parse_bare_specifier.rs
@@ -17,8 +17,8 @@ use derive_more::{Display, Error};
 use miette::Diagnostic;
 use node_semver::{Range, Version};
 use pnpm_network::percent_decode_str;
+use pnpm_package_name::is_valid_old_npm_package_name;
 use pnpm_resolving_jsr_specifier_parser::{ParseJsrSpecifierError, parse_jsr_specifier};
-use pnpm_resolving_parse_wanted_dependency::is_valid_old_npm_package_name;
 use pnpm_resolving_resolver_base::{
     ANY_VERSION_RANGE, is_any_version_range, is_valid_semver_range,
 };
