+++ pnpm/crates/resolving-parse-wanted-dependency/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace     = true
 repository.workspace  = true
 
 [dependencies]
+pnpm-package-name = { workspace = true }
 
 [lints]
 workspace = true
