+++ pnpm/crates/resolving-parse-wanted-dependency/src/lib.rs
@@ -2,7 +2,7 @@
 //! raw manifest or `add` argument into its `(alias, bareSpecifier)` halves and
 //! compares the supported equivalent forms of Git specifiers.
 
-pub mod validate_npm_package_name;
+pub use pnpm_package_name as validate_npm_package_name;
 
 pub use git_specifier::git_specifiers_are_equivalent;
 pub use validate_npm_package_name::is_valid_old_npm_package_name;
