+++ pnpm/crates/store-dir/Cargo.toml
@@ -11,11 +11,11 @@ license.workspace    = true
 repository.workspace = true
 
 [dependencies]
-pnpm-crypto-hash                       = { workspace = true }
-pnpm-deps-path                         = { workspace = true }
-pnpm-fs                                = { workspace = true }
-pnpm-resolving-parse-wanted-dependency = { workspace = true }
-pnpm-shared-artifact-protocol          = { workspace = true }
+pnpm-package-name             = { workspace = true }
+pnpm-crypto-hash              = { workspace = true }
+pnpm-deps-path                = { workspace = true }
+pnpm-fs                       = { workspace = true }
+pnpm-shared-artifact-protocol = { workspace = true }
 
 dashmap       = { workspace = true }
 derive_more   = { workspace = true }
